+++ cogeo_mosaic/backends/stac.py
@@ -53,7 +53,7 @@ class STACBackend(BaseBackend):
 
     """
 
-    path: str = attr.ib()
+    input: str = attr.ib()
     query: Dict = attr.ib()
     minzoom: int = attr.ib()
     maxzoom: int = attr.ib()
@@ -81,7 +81,7 @@ class STACBackend(BaseBackend):
     _backend_name = "STAC"
 
     def __attrs_post_init__(self):
-        """Post Init: if not passed in init, try to read from self.path."""
+        """Post Init: if not passed in init, try to read from self.input."""
         self.mosaic_def = self._read()
         self.bounds = self.mosaic_def.bounds
 
@@ -93,9 +93,9 @@ class STACBackend(BaseBackend):
             MosaicJSON: Mosaic definition.
 
         """
-        logger.debug(f"Using STAC backend: {self.path}")
+        logger.debug(f"Using STAC backend: {self.input}")
 
-        features = _fetch(self.path, self.query, **self.stac_api_options,)
+        features = _fetch(self.input, self.query, **self.stac_api_options,)
         logger.debug(f"Creating mosaic from {len(features)} features")
 
         # We need a specific accessor for STAC